fix: Fix issue with saving settings when navigating from listing
