+++ static/js/global.d.ts
@@ -15,26 +15,6 @@ declare interface Window {
   DNS_VERIFICATION_TOKEN: string;
   SENTRY_DSN: string;
   CSRF_TOKEN: string;
-  SNAP_SETTINGS_DATA: {
-    blacklist_countries: string[];
-    blacklist_country_keys: string;
-    countries: Array<{ key: string; name: string }>;
-    country_keys_status: string | null;
-    private: boolean;
-    publisher_name: string;
-    snap_id: string;
-    snap_name: string;
-    snap_title: string;
-    status: string;
-    store: string;
-    territory_distribution_status: string;
-    unlisted: boolean;
-    update_metadata_on_release: boolean;
-    visibility: string;
-    visibility_locked: boolean;
-    whitelist_countries: string[];
-    whitelist_country_keys: string;
-  };
   SNAP_LISTING_DATA: {
     DNS_VERIFICATION_TOKEN: string;
   };
