+++ webapp/publisher/snaps/settings_views.py
@@ -23,6 +23,61 @@ def get_settings_json(snap_name):
 
 
 @login_required
+def get_settings_data(snap_name):
+    snap_details = publisher_api.get_snap_info(snap_name, flask.session)
+
+    if "whitelist_country_codes" in snap_details:
+        whitelist_country_codes = (
+            snap_details["whitelist_country_codes"]
+            if len(snap_details["whitelist_country_codes"]) > 0
+            else []
+        )
+    else:
+        whitelist_country_codes = []
+
+    if "blacklist_country_codes" in snap_details:
+        blacklist_country_codes = (
+            snap_details["blacklist_country_codes"]
+            if len(snap_details["blacklist_country_codes"]) > 0
+            else []
+        )
+    else:
+        blacklist_country_codes = []
+
+    countries = []
+    for country in pycountry.countries:
+        countries.append({"key": country.alpha_2, "name": country.name})
+
+    is_on_lp = False
+    lp_snap = launchpad.get_snap_by_store_name(snap_name)
+    if lp_snap:
+        is_on_lp = True
+
+    context = {
+        "snap_name": snap_details["snap_name"],
+        "snap_title": snap_details["title"],
+        "snap_id": snap_details["snap_id"],
+        "publisher_name": snap_details["publisher"]["display-name"],
+        "license": snap_details["license"],
+        "private": snap_details["private"],
+        "unlisted": snap_details["unlisted"],
+        "countries": countries,
+        "whitelist_countries": whitelist_country_codes,
+        "blacklist_countries": blacklist_country_codes,
+        "store": snap_details["store"],
+        "keywords": snap_details["keywords"],
+        "status": snap_details["status"],
+        "is_on_lp": is_on_lp,
+        "update_metadata_on_release": snap_details[
+            "update_metadata_on_release"
+        ],
+        "visibility_locked": snap_details["visibility_locked"] | False,
+    }
+
+    return flask.jsonify({"success": True, "data": context})
+
+
+@login_required
 def get_settings(snap_name, return_json=False):
     snap_details = publisher_api.get_snap_info(snap_name, flask.session)
 
@@ -81,12 +136,7 @@ def get_settings(snap_name, return_json=False):
 
 
 @login_required
-def post_settings_json(snap_name):
-    return post_settings(snap_name, return_json=True)
-
-
-@login_required
-def post_settings(snap_name, return_json=False):
+def post_settings_data(snap_name):
     changes = None
     changed_fields = flask.request.form.get("changes")
 
@@ -104,8 +154,7 @@ def post_settings(snap_name, return_json=False):
                 response = publisher_api.snap_metadata(
                     snap_id, flask.session, body_json
                 )
-                if return_json:
-                    return flask.jsonify(response)
+                return flask.jsonify(response)
             except StoreApiResponseErrorList as api_response_error_list:
                 if api_response_error_list.status_code == 404:
                     return flask.abort(
@@ -184,9 +233,4 @@ def post_settings(snap_name, return_json=False):
                 "other_errors": other_errors,
             }
 
-            if return_json:
-                return flask.jsonify(context)
-
-            return flask.render_template("store/publisher.html", **context)
-
-    return flask.redirect(flask.url_for(".get_settings", snap_name=snap_name))
+            return flask.jsonify(context)
